requirements.txt
Flask==2.1.3
flask-restx==1.2.0
Flask-Cors==4.0.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
mysql-connector-python==8.1.0
Werkzeug==2.0.2
lxml==4.7.1
OpenDartReader==0.2.3
scikit-learn==1.3.2
keras==2.13.1
inpinitiFinance
tensorflow==2.13.0